// adapters/mtproto/src/mtproto/gen/api.d.ts
export namespace api {
// ...
  export type InputFileLocation<
    K extends keyof _InputFileLocation = keyof _InputFileLocation
  > = ToUnderscore<_InputFileLocation, K>;
  export type _InputFileLocation = {
    "inputFileLocation": {
      volume_id: bigint;                      // long
      local_id: number;                       // int
      secret: bigint;                         // long
      file_reference: Uint8Array;             // bytes
    },
    "inputEncryptedFileLocation": {
      id: bigint;                             // long
      access_hash: bigint;                    // long
    },
    "inputDocumentFileLocation": {
      id: bigint;                             // long
      access_hash: bigint;                    // long
      file_reference: Uint8Array;             // bytes
      thumb_size: string;                     // string
    },
    "inputSecureFileLocation": {
      id: bigint;                             // long
      access_hash: bigint;                    // long
    },
    "inputTakeoutFileLocation": {}
    "inputPhotoFileLocation": {
      id: bigint;                             // long
      access_hash: bigint;                    // long
      file_reference: Uint8Array;             // bytes
      thumb_size: string;                     // string
    },
    "inputPhotoLegacyFileLocation": {
      id: bigint;                             // long
      access_hash: bigint;                    // long
      file_reference: Uint8Array;             // bytes
      volume_id: bigint;                      // long
      local_id: number;                       // int
      secret: bigint;                         // long
    },
    "inputPeerPhotoFileLocation": {
      big?: true;                             // flags.0?true
      peer: api.InputPeer;                    // InputPeer
      photo_id: bigint;                       // long
    },
    "inputStickerSetThumb": {
      stickerset: api.InputStickerSet;        // InputStickerSet
      thumb_version: number;                  // int
    },
    "inputGroupCallStream": {
      call: api.InputGroupCall;               // InputGroupCall
      time_ms: bigint;                        // long
      scale: number;                          // int
      video_channel?: number;                 // flags.0?int
      video_quality?: number;                 // flags.0?int
    },
  };
// ...
  export type Config<
    K extends keyof _Config = keyof _Config
  > = ToUnderscore<_Config, K>;
  export type _Config = {
    "config": {
      default_p2p_contacts?: true;            // flags.3?true
      preload_featured_stickers?: true;       // flags.4?true
      revoke_pm_inbox?: true;                 // flags.6?true
      blocked_mode?: true;                    // flags.8?true
      force_try_ipv6?: true;                  // flags.14?true
      date: number;                           // int
      expires: number;                        // int
      test_mode: boolean;                     // Bool
      this_dc: number;                        // int
      dc_options: api.DcOption[];             // Vector<DcOption>
      dc_txt_domain_name: string;             // string
      chat_size_max: number;                  // int
      megagroup_size_max: number;             // int
      forwarded_count_max: number;            // int
      online_update_period_ms: number;        // int
      offline_blur_timeout_ms: number;        // int
      offline_idle_timeout_ms: number;        // int
      online_cloud_timeout_ms: number;        // int
      notify_cloud_delay_ms: number;          // int
      notify_default_delay_ms: number;        // int
      push_chat_period_ms: number;            // int
      push_chat_limit: number;                // int
      edit_time_limit: number;                // int
      revoke_time_limit: number;              // int
      revoke_pm_time_limit: number;           // int
      rating_e_decay: number;                 // int
      stickers_recent_limit: number;          // int
      channels_read_media_period: number;     // int
      tmp_sessions?: number;                  // flags.0?int
      call_receive_timeout_ms: number;        // int
      call_ring_timeout_ms: number;           // int
      call_connect_timeout_ms: number;        // int
      call_packet_timeout_ms: number;         // int
      me_url_prefix: string;                  // string
      autoupdate_url_prefix?: string;         // flags.7?string
      gif_search_username?: string;           // flags.9?string
      venue_search_username?: string;         // flags.10?string
      img_search_username?: string;           // flags.11?string
      static_maps_provider?: string;          // flags.12?string
      caption_length_max: number;             // int
      message_length_max: number;             // int
      webfile_dc_id: number;                  // int
      suggested_lang_code?: string;           // flags.2?string
      lang_pack_version?: number;             // flags.2?int
      base_lang_pack_version?: number;        // flags.2?int
      reactions_default?: api.Reaction;       // flags.15?Reaction
      autologin_token?: string;               // flags.16?string
    },
  };
// ...
  export type InputWebFileLocation<
    K extends keyof _InputWebFileLocation = keyof _InputWebFileLocation
  > = ToUnderscore<_InputWebFileLocation, K>;
  export type _InputWebFileLocation = {
    "inputWebFileLocation": {
      url: string;                            // string
      access_hash: bigint;                    // long
    },
    "inputWebFileGeoPointLocation": {
      geo_point: api.InputGeoPoint;           // InputGeoPoint
      access_hash: bigint;                    // long
      w: number;                              // int
      h: number;                              // int
      zoom: number;                           // int
      scale: number;                          // int
    },
    "inputWebFileAudioAlbumThumbLocation": {
      small?: true;                           // flags.2?true
      document?: api.InputDocument;           // flags.0?InputDocument
      title?: string;                         // flags.1?string
      performer?: string;                     // flags.1?string
    },
  };
// ...
  export type FileHash<
    K extends keyof _FileHash = keyof _FileHash
  > = ToUnderscore<_FileHash, K>;
  export type _FileHash = {
    "fileHash": {
      offset: bigint;                         // long
      limit: number;                          // int
      hash: Uint8Array;                       // bytes
    },
  };
// ...
}

export namespace storage {
  export type FileType<
    K extends keyof _FileType = keyof _FileType
  > = ToUnderscore<_FileType, K>;
  export type _FileType = {
    "storage.fileUnknown": {}
    "storage.filePartial": {}
    "storage.fileJpeg": {}
    "storage.fileGif": {}
    "storage.filePng": {}
    "storage.filePdf": {}
    "storage.fileMp3": {}
    "storage.fileMov": {}
    "storage.fileMp4": {}
    "storage.fileWebp": {}
  };

  export const fileUnknown: TLConstructorEmpty<"storage.fileUnknown">;
  export const filePartial: TLConstructorEmpty<"storage.filePartial">;
  export const fileJpeg: TLConstructorEmpty<"storage.fileJpeg">;
  export const fileGif: TLConstructorEmpty<"storage.fileGif">;
  export const filePng: TLConstructorEmpty<"storage.filePng">;
  export const filePdf: TLConstructorEmpty<"storage.filePdf">;
  export const fileMp3: TLConstructorEmpty<"storage.fileMp3">;
  export const fileMov: TLConstructorEmpty<"storage.fileMov">;
  export const fileMp4: TLConstructorEmpty<"storage.fileMp4">;
  export const fileWebp: TLConstructorEmpty<"storage.fileWebp">;
}
// ...
export namespace upload {
  export type File<
    K extends keyof _File = keyof _File
  > = ToUnderscore<_File, K>;
  export type _File = {
    "upload.file": {
      type: storage.FileType;                 // storage.FileType
      mtime: number;                          // int
      bytes: Uint8Array;                      // bytes
    },
    "upload.fileCdnRedirect": {
      dc_id: number;                          // int
      file_token: Uint8Array;                 // bytes
      encryption_key: Uint8Array;             // bytes
      encryption_iv: Uint8Array;              // bytes
      file_hashes: api.FileHash[];            // Vector<FileHash>
    },
  };

  export const file: TLConstructor<_File, "upload.file">;
  export const fileCdnRedirect: TLConstructor<_File, "upload.fileCdnRedirect">;
  export type WebFile<
    K extends keyof _WebFile = keyof _WebFile
  > = ToUnderscore<_WebFile, K>;
  export type _WebFile = {
    "upload.webFile": {
      size: number;                           // int
      mime_type: string;                      // string
      file_type: storage.FileType;            // storage.FileType
      mtime: number;                          // int
      bytes: Uint8Array;                      // bytes
    },
  };

  export const webFile: TLConstructor<_WebFile, "upload.webFile">;
  export type CdnFile<
    K extends keyof _CdnFile = keyof _CdnFile
  > = ToUnderscore<_CdnFile, K>;
  export type _CdnFile = {
    "upload.cdnFileReuploadNeeded": {
      request_token: Uint8Array;              // bytes
    },
    "upload.cdnFile": {
      bytes: Uint8Array;                      // bytes
    },
  };

  export const cdnFileReuploadNeeded: TLConstructor<_CdnFile, "upload.cdnFileReuploadNeeded">;
  export const cdnFile: TLConstructor<_CdnFile, "upload.cdnFile">;
}
// ...
export namespace upload {
  export const saveFilePart: TLApiMethod<"upload.saveFilePart", {
    file_id: bigint                         // long
    file_part: number                       // int
    bytes: Uint8Array                       // bytes
  }, boolean>
  export const getFile: TLApiMethod<"upload.getFile", {
    precise?: true                          // flags.0?true
    cdn_supported?: true                    // flags.1?true
    location: api.InputFileLocation         // InputFileLocation
    offset: bigint                          // long
    limit: number                           // int
  }, File>
  export const saveBigFilePart: TLApiMethod<"upload.saveBigFilePart", {
    file_id: bigint                         // long
    file_part: number                       // int
    file_total_parts: number                // int
    bytes: Uint8Array                       // bytes
  }, boolean>
  export const getWebFile: TLApiMethod<"upload.getWebFile", {
    location: api.InputWebFileLocation      // InputWebFileLocation
    offset: number                          // int
    limit: number                           // int
  }, WebFile>
  export const getCdnFile: TLApiMethod<"upload.getCdnFile", {
    file_token: Uint8Array                  // bytes
    offset: bigint                          // long
    limit: number                           // int
  }, CdnFile>
  export const reuploadCdnFile: TLApiMethod<"upload.reuploadCdnFile", {
    file_token: Uint8Array                  // bytes
    request_token: Uint8Array               // bytes
  }, api.FileHash[]>
  export const getCdnFileHashes: TLApiMethod<"upload.getCdnFileHashes", {
    file_token: Uint8Array                  // bytes
    offset: bigint                          // long
  }, api.FileHash[]>
  export const getFileHashes: TLApiMethod<"upload.getFileHashes", {
    location: api.InputFileLocation         // InputFileLocation
    offset: bigint                          // long
  }, api.FileHash[]>
}